backend/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
cherche==2.2.1
Django==4.2.16
django-cors-headers==4.4.0
djangorestframework==3.15.2
drf-yasg==1.21.7
filelock==3.16.1
flashtext==2.7
fsspec==2024.9.0
gunicorn==23.0.0
huggingface-hub==0.25.0
idna==3.10
inflection==0.5.1
Jinja2==3.1.4
joblib==1.4.2
lenlp==1.1.0
lunr==0.7.0.post1
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
packaging==24.1
pandas==2.2.2
pillow==10.4.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
rapidfuzz==3.9.7
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.1.0
setuptools==75.1.0
six==1.16.0
sqlparse==0.5.1
sympy==1.13.2
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.4.1
tqdm==4.66.5
transformers==4.44.2
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
